backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
Faker==23.2.1
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==23.2
pep8-naming==0.13.3
pillow==10.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
